\name{electionAPI-package}
\alias{electionAPI-package}
\alias{electionAPI}
\docType{package}
\title{
\packageTitle{electionAPI}
}
\description{
\packageDescription{electionAPI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{electionAPI}
\packageIndices{electionAPI}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{electionAPI}

Maintainer: \packageMaintainer{electionAPI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
